vareeprom_common.sh: Add support for EEPROM offsets > 255
Currently, the various *_vareeprom_dram programming scripts are limited
to EEPROM data addresses of 255 or less due to their use of i2c-tools.
This limits reading and writing data to the first page of EEPROM.
However, EEPROMs sized 2 KB and under actually reserve up to 3 of the
LSBs of the 7-Bit I2C address as page address bits to allow addressing
of additional 256 byte pages.

Therefore, translate the address and offset and pass to the i2c
functions in order to support offsets greater than 255. This will allow
support for EEPROMs up to 2K - beyond this, additional functionality
would be needed as larger EEPROMs do not use page address bits but have
word addresses greater than 1 byte specified in the I2C data cycle.

Signed-off-by: Ken Sloat <ken.s@variscite.com>