var_kit_test: increase udhcpc attempts from 3 to 10
The script currently sleeps 3 seconds after starting wpa_supplicant.
This may not always be enough time to associate with the access point
and be ready to dhcp.

Increasing the udhcpc attempts from 3 to 10 adds additional time before
udhcpc will give up. It's still an arbitrary number, but this is more
robust and deterministic than increasing the sleep time.

Signed-off-by: Nate Drude <nate.d@variscite.com>